Monitoring plugin to separate directory
