--- yandextank/monitoring/agent/__init__.py
@@ -1,4 +1,0 @@
-'''
-Agent to be installed at remote server
-'''
-#TODO: 2 implement single process monitoring metric
