+++ yandextank/plugins/Monitoring/__init__.py
@@ -2,3 +2,4 @@
 Package contains all original tank tool plugins
 '''
 from collector import *
+from plugin import *
