+++ yandextank/plugins/Monitoring/agent/__init__.py
@@ -1,0 +1,4 @@
+'''
+Agent to be installed at remote server
+'''
+#TODO: 2 implement single process monitoring metric
