+++ yandextank/plugins/Monitoring/collector.py
